Create Nifi transform for requisition line items
This requires some additional logic in the requisition transform. Requisitions are saved in a cache with the V3 UUID mapping, in order to ensure coordination (requisitions saved before line items to avoid foreign key violations). This is saved after publishing to Kafka, in order for requisitions to go through the pipeline before line items. Each line item checks for existence of its parent requisition in the cache before continuing through the pipeline. Line items are also saved in the cache for their children (losses and adjustments).

Redis with persistence and a mapping volume has been added to docker compose for this cache.